Refactoring: total separation between UI code and network/issuing code
All network and issuing code now runs on its own separate thread, and this thread is managed by the EnrollClient class. This class offers a single method, EnrollClient.enroll(), that gets the process going, and when done it reports back to the calling thread via a specified Handler.

The following has also changed:
* we now connect the MNO after we've made contact with the passport, instead of before
* we no longer connect to the government server; for now we just issue the credential directly