Don't close dialog when showing the disclosure info activity
The "you don't have the right attributes"-dialog now remains open
in the background while the user sees the disclosure info activity.
For symmetry with the dialog that the user sees when he does have
the right attributes (which also has a more info-button, and which
also remains in the background). Also improves the workflow when
we were called from an intent from the mobile browser